+++ weaver/processes/wps_package.py
@@ -1234,8 +1234,14 @@ class DirectoryNestedStorage(CachedStorage):
 
 
 class WpsPackage(Process):
-    def __init__(self, package=None, payload=None, **kw):
-        # type: (CWL, Optional[JSON], **Any) -> None
+    def __init__(
+        self,
+        *,
+        package=None,   # type: CWL
+        payload=None,   # type; Optional[JSON]
+        settings=None,  # type: Optional[AnySettingsContainer]
+        **kw,           # type: Any
+    ):                  # type: (...) -> None
         """
         Creates a `WPS-3 Process` instance to execute a `CWL` application package definition.
 
@@ -1264,7 +1270,7 @@ class WpsPackage(Process):
 
         self.payload = payload
         self.package = package
-        self.settings = get_settings()
+        self.settings = get_settings(settings)
         if not self.package:
             raise PackageRegistrationError("Missing required package definition for package process.")
         if not isinstance(self.package, dict):